--- taskqueue/gcloud/aio/taskqueue/error.py
@@ -1,2 +1,0 @@
-class FailFastError(Exception):
-    pass
