--- taskqueue/tests/unit/error_test.py
@@ -1,5 +1,0 @@
-import gcloud.aio.taskqueue.error as error  # pylint: disable=unused-import
-
-
-def test_importable():
-    assert True
