--- taskqueue/tests/unit/pullqueue_test.py
@@ -1,5 +1,0 @@
-import gcloud.aio.taskqueue.pullqueue as pullqueue  # pylint: disable=unused-import
-
-
-def test_importable():
-    assert True
