--- taskqueue/tests/unit/taskmanager_test.py
@@ -1,5 +1,0 @@
-import gcloud.aio.taskqueue.taskmanager as taskmanager  # pylint: disable=unused-import
-
-
-def test_importable():
-    assert True
